+++ packages/storage/__tests__/providers/s3/apis/uploadData/putObjectJob.test.ts
@@ -5,7 +5,7 @@ import { Credentials } from '@aws-sdk/types';
 import { Amplify } from '@aws-amplify/core';
 import { putObject } from '../../../../../src/providers/s3/utils/client';
 import { calculateContentMd5 } from '../../../../../src/providers/s3/utils';
-import { putObjectJob } from '../../../../../src/providers/s3/apis/uploadData/putObjectJob';
+import { putObjectJob } from '../../../../../src/providers/s3/apis/upload/putObjectJob';
 
 jest.mock('../../../../../src/providers/s3/utils/client');
 jest.mock('../../../../../src/providers/s3/utils', () => {
