+++ packages/storage/src/providers/s3/apis/upload/multipart/uploadHandlers.ts
@@ -4,7 +4,7 @@
 import { Amplify, StorageAccessLevel } from '@aws-amplify/core';
 
 import { getDataChunker } from './getDataChunker';
-import { UploadDataInput } from '../../../types';
+import { UploadInput } from '../../../types';
 import { resolveS3ConfigAndInput } from '../../../utils';
 import { Item as S3Item } from '../../../types/outputs';
 import {
@@ -32,7 +32,7 @@ import {
  * @internal
  */
 export const getMultipartUploadHandlers = (
-	{ options: uploadDataOptions, key, data }: UploadDataInput,
+	{ options: uploadDataOptions, key, data }: UploadInput,
 	size?: number
 ) => {
 	let resolveCallback: ((value: S3Item) => void) | undefined;
